Write hero section text styles
I had to modify the font URL so it uses the correct font for bold text.
Not that it was absolutely necessary, but now I think I know why my text
looks off when compared with an original.

I set the hero font size as a variable in the hero styles so I can keep
the media query short. I think I'll like this style of writing media
queries, but it will probably get cumbersome and confusing quickly, when
a lot of style changes are involved.

I also had to sneak in changing the narrow container's max width so the
way the text looks is identical to the original.